\name{nonmolar}
\alias{nonmolar}
\docType{data}
\title{Non-molar tooth loss dataset.}
\description{
The non-molar tooth loss dataset is from Teeth, which is contained in R Package MST. Teeth consists of 4958 patients with 40196 non-molar teeth. For each non-molar tooth, there are 52 covariates.
}
\usage{data("nonmolar")}
\format{A data.frame containing 40196 observations from 4958 patients.}
\references{
Calhoun, Peter and Su, Xiaogang and Nunn, Martha and Fan, Juanjuan (2018) \emph{Constructing Multivariate Survival Trees: The MST Package for R}. Journal of Statistical Software.
}
\keyword{datasets}